requirements.txt
absl-py==1.4.0
asttokens==2.2.1
backcall==0.2.0
boto3==1.26.114
botocore==1.29.114
cachetools==5.3.0
certifi==2022.12.7
chardet==4.0.0
charset-normalizer==3.1.0
click==8.0.4
colorama==0.4.6
contourpy==1.0.7
cycler==0.10.0
decorator==5.1.1
executing==1.2.0
filelock==3.11.0
fire==0.5.0
Flask==2.2.3
fonttools==4.39.3
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.17.3
google-auth-oauthlib==1.0.0
grpcio==1.53.0
huggingface-hub==0.13.4
idna==2.10
ipython==8.12.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jmespath==1.0.1
kiwisolver==1.4.4
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mpmath==1.3.0
networkx==3.1
numpy==1.24.2
oauthlib==3.2.2
opencv-python==4.7.0.72
packaging==23.1
pandas==2.0.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.5.0
prompt-toolkit==3.0.38
protobuf==4.22.3
psutil==5.9.4
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybboxes==0.1.6
Pygments==2.15.0
pyparsing==2.4.7
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
roboflow==1.0.5
rsa==4.9
s3transfer==0.6.0
sahi==0.11.13
scipy==1.10.1
seaborn==0.12.2
shapely==2.0.1
six==1.16.0
smmap==5.0.0
stack-data==0.6.2
sympy==1.11.1
tensorboard==2.12.2
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
termcolor==2.2.0
terminaltables==3.1.10
thop==0.1.1.post2209072238
torch==2.0.0
torchvision==0.15.1
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
wcwidth==0.2.6
Werkzeug==2.2.3
wget==3.2
yolov5==7.0.11
